Combined existing modules with redis
