+++ constants.py
@@ -1,0 +1,8 @@
+REDIS_DB_FILE = '/tmp/redis.db'
+
+# REDIS CHANNEL NAMES
+UI_CHANNEL = b'ui'
+EMOTION_CHANNEL = b'emotion'
+SPEECH_CHANNEL = b'speech'
+TTS_CHANNEL = b'tts'
+LEDS_CHANNEL = b'leds'
