+++ ui/templates/base.html
@@ -48,9 +48,13 @@
         <script src="{{ url_for('static', filename='js/bootstrap.min.js') }}"></script>
         <script src="//cdnjs.cloudflare.com/ajax/libs/socket.io/2.2.0/socket.io.js" integrity="sha256-yr4fRk/GU1ehYJPAs8P4JlTgu0Hdsp4ZKrx8bDEDC3I=" crossorigin="anonymous"></script>
         <script type="text/javascript" charset="utf-8">
-            var socket = io();
+            var socket = io.connect('/ui');
+            console.log("Socket object: " + socket)
+            socket.on('connect', () => {
+                console.log("Connected. ID: " + socket.id);
+            });
             socket.on('ui', function(data) {
-                console.log(data);
+                console.log('Received data from server: ' + JSON.stringify(data));
             });
         </script>
     </body>
